% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{checkutilC}
\alias{checkutilC}
\title{Internal function to check if the observed recruitment tree is returned by any
  given Ustar and Vstar matrices.}
\usage{
checkutilC(Ustar, Vstar, netstage, recfromUV3urR, reslocal)
}
\arguments{
\item{Ustar}{Matrix of utilities from recruiters to peers.}

\item{Vstar}{Matrix of utilities from peers to recruiters.}

\item{netstage}{Amended adjacency matrix, with observed recruitment information.}

\item{recfromUV3urR}{Call to another \code{C++} function.}

\item{reslocal}{Logical: use additional information about the underlying network?
Makes computation faster. If TRUE, assumes that \code{netstage} takes three values:
1 indicates a recruitment tie (and thus a tie in the underlying network); 0 indicates
no recruitment tie, but a tie does exist in the underlying network; NA indicates no
tie in the underlying network, and thus no possibility of recruitment.}
}
\value{
TRUE or FALSE if the given Ustar and Vstar return the observed recruitment tree.
}
\description{
Internal function to check if the observed recruitment tree is returned by any
  given Ustar and Vstar matrices.
}
