% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumfuncs.R
\name{dyadmatch}
\alias{dyadmatch}
\title{Tabulate dyad matches.}
\usage{
dyadmatch(rdf)
}
\arguments{
\item{rdf}{An object of class \code{rds.data.frame}, with one column for the characteristic of interest, named char.}
}
\value{
Table with dyad match possibilities.
}
\description{
Tabulate if both members of the dyad have the same characteristic value.
}
