% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumfuncs.R
\name{dyadtype}
\alias{dyadtype}
\title{Return table of all dyad types}
\usage{
dyadtype(rdf)
}
\arguments{
\item{rdf}{An object of class \code{rds.data.frame}, with one column for the characteristic of interest, named char.}
}
\value{
A table with all dyad match types
}
\description{
Return table of all dyad types
}
