% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{firstZero}
\alias{firstZero}
\title{Internal (cpp) of matching function: finds first zero in a vector.}
\usage{
firstZero(x)
}
\arguments{
\item{x}{An integer vector.}
}
\value{
Position in vector (R-vector; cannot be 0) where first zero occurs.
  Returns 0 if there are no zeros.
}
\description{
Internal (cpp) of matching function: finds first zero in a vector.
}
