% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gennewchain.R
\name{gennewchain}
\alias{gennewchain}
\title{Generate a new recruitment chain from posterior medians.}
\usage{
gennewchain(alphaout, betaout, usenet = FALSE, xiout, zetaout, rdsdf, covname)
}
\arguments{
\item{alphaout}{Vector of \code{alpha} values.}

\item{betaout}{Vector of \code{beta} values.}

\item{usenet}{Logical: use the network size variable? Defaults to false.}

\item{xiout}{Vector of \code{xi} values.}

\item{zetaout}{Numerical value of \code{zeta}.}

\item{rdsdf}{An rds.data.frame object containing the full recruitment chain.}

\item{covname}{The name of the covariate to be used in the model.}
}
\value{
The outcome of a new \code{simrec} call, which is a list consisting of
 \item{net}{network object for the full recruitment chain}
 \item{rds}{rds.data.frame object for the full recruitment chain}
 \item{rdsfull}{rds.data.frame object including population members not in the sample. For simulation studies. These members of pop not in the sample have recruiter value NA, wave value 1000 (larger than would be encountered).}
 \item{seedsrec}{indicator of whether or not the seeds recruited anyone}
 \item{alphaorig}{original value of alpha used for simulation}
 \item{betaorig}{original value of beta used for simulation}
}
\description{
Use particular values of \code{alpha}, \code{beta}, \code{xi}, and \code{zeta} (e.g. resulting from fitting an RCPS model) to generate a new recruitment chain with the same seeds as a given RDS sample.
}
