% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbsC}
\alias{gibbsC}
\title{Internal (cpp) function for the Gibbs sampler iterations.}
\usage{
gibbsC(
  Kcalc,
  alphastart,
  betastart,
  adjmat,
  nrec,
  npeer,
  X,
  Y,
  Ustar,
  Vstar,
  expgrid,
  rtruncnorm,
  sample,
  ma0,
  sa0,
  mb0,
  sb0,
  tallowed,
  recfromUV3urR,
  reslocal,
  mvrnorm,
  updlast
)
}
\arguments{
\item{Kcalc}{The number of iterations to run.  Equal to desired number
of draws times interval, plus the burn-in period.}

\item{alphastart}{Initialization value for alpha, preference coefficient for recruiters.}

\item{betastart}{Initialization value for beta, preference coefficient for peers.}

\item{adjmat}{List of amended adjacency matrices, one for each wave.  Each contains
recruitment information for this wave plus information about self-matches. Assumed to
take two values (0-1) if \code{restrict.local=FALSE}, or three values (0-1-NA) if
\code{restrict.local=TRUE}.}

\item{nrec}{List of number of recruiters, one value for each wave.}

\item{npeer}{List of number of peers, one value for each wave.}

\item{X}{List of list of covariate(s) matches from recruiters to peers.}

\item{Y}{List of list of covariate(s) matches from peers to recruiters.}

\item{Ustar}{List of matrices of utilities from recruiters to peers, one for each wave.}

\item{Vstar}{List of matrices of utilities from peers to recruiters, one for each wave.}

\item{expgrid}{R function \code{expand.grid} from \code{base}.}

\item{rtruncnorm}{R function \code{rtruncnorm} from the \code{truncnorm} package.}

\item{sample}{R function \code{sample} from \code{base}.}

\item{ma0}{Prior value \eqn{\mu_\alpha}, mean of normal distribution for alpha.}

\item{sa0}{Prior value \eqn{\Sigma_\alpha}, variance of normal distribution for alpha.}

\item{mb0}{Prior value \eqn{\mu_\beta}, mean of normal distribution for beta.}

\item{sb0}{Prior value \eqn{\Sigma_\beta}, varaince of normal distribution for beta.}

\item{tallowed}{Number of hours allowed on Hoffman.}

\item{recfromUV3urR}{R function in this package.}

\item{reslocal}{Logical: use additional information about the underlying network? Makes
computation faster. If TRUE, assumes that \code{netstage} takes three values: 1 indicates a
recruitment tie (and thus a tie in the underlying network); 0 indicates no recruitment tie,
but a tie does exist in the underlying network; NA indicates no tie in the underlying network,
and thus no possibility of recruitment.}

\item{mvrnorm}{Function to make multivariate normal draws.}

\item{updlast}{Logical: flag to stop. Not currently used.}
}
\value{
A list containing
  \item{alpha}{Draws of alpha from the posterior distribution.}
  \item{beta}{Draws of beta from the posterior distribution.}
  \item{rejcounter}{Number of rejections (bc of swapping problem).}
  \item{earlystop}{Number of iterations after which Gibbs sampler was forced to abort for time constraints.}
}
\description{
Setup must already have occurred.
}
