% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{listReassign}
\alias{listReassign}
\title{Internal (cpp) of matching function: housekeeping.}
\usage{
listReassign(nrec, npeer, jchoice, RS)
}
\arguments{
\item{nrec}{Number of recruiters.}

\item{npeer}{Number of peers.}

\item{jchoice}{Vector of length \code{npeer} with each peer's selection
(can be self).}

\item{RS}{List.}
}
\value{
A list with the new recruiter selections.
}
\description{
Function that carries out list reassignment: takes the vector of peer
  choices at a given stage of matching and makes a new selection list.
}
