% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makenetstageR.R
\name{makeNetStageR}
\alias{makeNetStageR}
\title{Make annotated adjacency matrix for a given wave.}
\usage{
makeNetStageR(
  rdsdf,
  covnames,
  wave,
  restrict.local = FALSE,
  popadjmat = NULL,
  usenetsize = FALSE
)
}
\arguments{
\item{rdsdf}{An rds.data.frame object containing the full recruitment chain.  Must also contain information about covariate(s) for each person.}

\item{covnames}{Vector of names of covariates in \code{rdf} to be used in the model.}

\item{wave}{The wave of the recruiter, so choose 0 for wave where seeds recruit wave 1.}

\item{restrict.local}{Logical: use additional information about the underlying network? Makes computation faster.  Must also specify \code{popadjmat}.}

\item{popadjmat}{Population adjacency matrix.  Must be supplied if \code{restrict.local=TRUE}.}

\item{usenetsize}{Logical: use network size? Defaults to false.}
}
\value{
A list containing
 \item{netstage}{the annotated adjacency matrix, containing observed recruitment information and self-match information for the given wave.  Assumed to take two values (0-1) if \code{restrict.local=FALSE}, or three values (0-1-NA) if \code{restrict.local=TRUE}.}
 \item{chrec}{the covariate values for this wave's recruiters}
 \item{chpeer}{the covariate values for this wave's peers}
}
\description{
Create a "netstage" (amended adjacency matrix) matrix for a given wave of recruitment.  For \code{n_r} recruiters and \code{n_p} peers in wave \code{w}, this is a \eqn{(n_r+1) \times (n_p+3)} matrix, where the final row and final three columns contain self-match information.  Should be called in a urInfR run.
}
