% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{makeU}
\alias{makeU}
\title{Internal funcion to construct U given alpha, X, epsilon}
\usage{
makeU(charrec, charpeer, alpha, ksi, epsilon)
}
\arguments{
\item{charrec}{Matrix of covariate values for recruiters.}

\item{charpeer}{Matrix of covariate values for peers.}

\item{alpha}{Preference coefficients for recruiters to peers.}

\item{ksi}{Parameter for number of recruits. Vector length ncoupon.}

\item{epsilon}{Matrix with random error.  Must have same dimension as U.}
}
\value{
Matrix with calculated values of \eqn{U}, where \eqn{U =\alpha X + \epsilon}
}
\description{
Note that this is the CALCULATION, only used in step 0
}
