% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{makeV}
\alias{makeV}
\title{Internal funcion to construct V given beta, Y, gamma}
\usage{
makeV(charrec, charpeer, beta, zeta, gamma)
}
\arguments{
\item{charrec}{Matrix of covariate values for recruiters.}

\item{charpeer}{Matrix of covariate values for peers.}

\item{beta}{Preference coefficients for peers to recruiters. Last value is self.}

\item{zeta}{Parameter for self match. Vector length 1.}

\item{gamma}{Matrix with random error.  Must have same dimension as V.}
}
\value{
Matrix with calculated values of \eqn{V}, where \eqn{V =\beta Y + \gamma}
}
\description{
Note that this is the CALCULATION, only used in step 0
}
