% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumfuncs.R
\name{numRecCount}
\alias{numRecCount}
\title{Returns a table showing how many recruiters enrolled 0-3 people}
\usage{
numRecCount(netstage)
}
\arguments{
\item{netstage}{An amended adjacency matrix for a particular wave.  The final row and final column must represent self-matches.}
}
\value{
A table showing how many recruiters enrolled 0, 1, 2, or 3 people into the study.
}
\description{
Returns a table showing how many recruiters enrolled 0-3 people
}
