% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{peerChoice}
\alias{peerChoice}
\title{Internal (cpp) of matching function: peer choices.}
\usage{
peerChoice(recsels, Vrank)
}
\arguments{
\item{recsels}{List with the (up to 3) selections of each recruiter.}

\item{Vrank}{matrix of ranks from each peer to each recruiter and self.}
}
\value{
A vector with the index of each peer's choice.
}
\description{
For each peer, returns their choice from amongst the recruiters
  that selected them at this stage and themself. Will always return
  a value (self), even if not selected.
}
