% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RaddEL.R
\name{RaddEL}
\alias{RaddEL}
\title{Generate recruitment edgelist for a specific wave.}
\usage{
RaddEL(
  w,
  N,
  alpha,
  beta,
  ksi,
  zeta,
  elbuild,
  pindex,
  ch,
  usenetsize,
  netsiz,
  netr,
  netp
)
}
\arguments{
\item{w}{Wave number.}

\item{N}{Population size.}

\item{alpha}{Preference coefficient for recruiters to peers.}

\item{beta}{Preference coefficient for peers to recruiters.}

\item{ksi}{Preference coefficient for abstention.}

\item{zeta}{Preference coefficient for declension.}

\item{elbuild}{Existing edgelist (for waves 0 to \code{w}-1).}

\item{pindex}{Index of peers for wave \code{w}.}

\item{ch}{Covariate(s).  Must have length \code{N}.}

\item{usenetsize}{Logical: use network size?}

\item{netsiz}{Vector of network sizes. Must be provided if \code{usenetsize} is true.}

\item{netr}{Must be provided if \code{usenetsize} is true.}

\item{netp}{Must be provided if \code{usenetsize} is true.}
}
\value{
Edgelist for wave \code{w}.
}
\description{
Simulation mechanism: generate recruitment edgelist for a specific wave.  Internal function called by sim_mecC.R
}
