% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumfuncs.R
\name{rec.homoph}
\alias{rec.homoph}
\title{Return information about recruitment homophily, with parallel option.}
\usage{
rec.homoph(simout, vname = "char")
}
\arguments{
\item{simout}{The output of a simrec function call in \code{prefrecruit}. Must have an object of class \code{rds.data.frame} called \code{rds}, with one column for the characteristic of interest, named char.}

\item{vname}{Name of outcome variable.}
}
\value{
The recruitment homophily and p-value for a chi-square test of independence.
}
\description{
For multiple chains, use \code{sapply(listname, rec.homoph)}.
}
