% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_thrC.R
\name{recfromUV3urR}
\alias{recfromUV3urR}
\title{Generate recruitment chain based on utilities.}
\usage{
recfromUV3urR(U, V, reslocal = FALSE, siml = FALSE)
}
\arguments{
\item{U}{Matrix of utilities from recruiters to peers.  It has dimension \eqn{n_r \times (n_p + n_c)}. Currently implemented for \eqn{n_c=3}.}

\item{V}{Matrix of utilities from peers to recruiters.  It has dimension \eqn{n_p \times (n_r + 1)}.}

\item{reslocal}{Logical: use additional information about the underlying network? Makes
computation faster. If TRUE, assumes that \code{netstage} takes three values: 1 indicates a
recruitment tie (and thus a tie in the underlying network); 0 indicates no recruitment tie,
but a tie does exist in the underlying network; NA indicates no tie in the underlying network,
and thus no possibility of recruitment.}

\item{siml}{Logical: is this a simulation (gives warning when < 3 peers).}
}
\value{
An annotated adjacency matrix with the generated recruitment chain.  It has dimension \eqn{(n_r+n_c) \times (n_p+1)} where \eqn{n_r} is the number of recruiters, \eqn{n_p} is the number of peers, and \eqn{n_c} is the number of coupons.
}
\description{
Generates a recruitment chain based on the given utilities \code{U} and \code{V}.  Uses the multiple matching process I developed.  This is the matching algorithm for the UNRANKED process (i.e., if a person recruits three peers, we do not make assumptions about those peers' relative utilities.)
}
\examples{
U <- matrix(rnorm(50), nrow=5)
V <- matrix(rnorm(42), nrow=7)
recfromUV3urR(U, V)
}
