% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtpbox.R
\name{rtpbox}
\alias{rtpbox}
\title{Reingold Tilford plot with each recruitment chain placed in a separate box.}
\usage{
rtpbox(rdf, char = "char", title = "Recruitment Plot")
}
\arguments{
\item{rdf}{An object of class \code{rds.data.frame}, with a column named \code{char} that contains node categories.}

\item{char}{Column name on \code{rdf} used to color the points on the plot.}

\item{title}{Title of plot.}
}
\value{
A recruitment plot.
}
\description{
This function also adds specificity for networks generated as part of the preferential recruitment modeling framework
}
\examples{
nseeds <- 3
N <- 2000
char <- matrix(sample(0:1, N, replace=TRUE), ncol=1, dimnames=list(1:N,"ch1"))
alpha <- 0
beta <- 0
ksi <- c(5,3,0)
zeta <- 1
waves <- 7

a <- simrec(nseeds, N, char, alpha, beta, ksi, zeta, waves)
rtpbox(a$rds, char="ch1")
}
