% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mecC.R
\name{simrec}
\alias{simrec}
\title{Simulate recruitment chains given parameters.}
\usage{
simrec(
  nseeds,
  N,
  ch,
  alpha,
  beta,
  ksi,
  zeta,
  waves,
  ncoup = 3,
  restrict.local = FALSE,
  popadjmat = NULL,
  specifyseeds = FALSE,
  seedID = NULL,
  targetn = FALSE,
  tss = NULL,
  usenetsize = FALSE,
  netsiz = NULL,
  netr = NULL,
  netp = NULL,
  outcome = NULL
)
}
\arguments{
\item{nseeds}{Number of seeds.}

\item{N}{Population size (including seeds).}

\item{ch}{Matrix of covariate(s).  Must have \code{N} rows.}

\item{alpha}{Preference coefficients for recruiters to peers. The last value is for self-match.}

\item{beta}{Preference coefficients for peers to recruiters. The last value is for self-match.}

\item{ksi}{Number of recruits parameter.}

\item{zeta}{Peer self-match parameter.}

\item{waves}{Number of waves.}

\item{ncoup}{Maximum number of coupons; set to 3 by default.}

\item{restrict.local}{Logical: use additional information about the underlying network? Makes computation faster. Must also specify \code{popadjmat}.}

\item{popadjmat}{Population adjacency matrix.  Must be supplied if \code{restrict.local=TRUE}.}

\item{specifyseeds}{Logical: use specific nodes as the seeds? Used for simulating inclusion probabilities after inference.  Must also specify \code{seedID}.}

\item{seedID}{ID of specific seeds to use if \code{specifyseeds=TRUE}.}

\item{targetn}{Logical: specify target sample size? FALSE by default. If true need to specify value to use n instead of number of waves as stopping point.}

\item{tss}{Target sample size}

\item{usenetsize}{Logical: use network size variable in model?}

\item{netsiz}{Network size of each member of the population. These will then be divided by the max within each set of recruiters and peers.}

\item{netr}{Preference coefficient recruiters have for large network sizes.}

\item{netp}{Preference coefficient peers have for large network sizes.}

\item{outcome}{Outcome measure of interest (e.g., HIV, HCV).}
}
\value{
A list consisting of
 \item{net}{network object for the full recruitment chain}
 \item{rds}{rds.data.frame object for the full recruitment chain}
 \item{rdsfull}{rds.data.frame object including population members not in the sample. For simulation studies. These members of pop not in the sample have recruiter value NA, wave value 1000 (larger than would be encountered).}
 \item{seedsrec}{indicator of whether or not the seeds recruited anyone}
 \item{alphaorig}{original value of alpha used for simulation}
 \item{betaorig}{original value of beta used for simulation}
}
\description{
Simulate recruitment chains given parameters.
}
\examples{
nseeds <- 2
N <- 500
char <- matrix(sample(0:1, N, replace=TRUE), ncol=1, dimnames=list(1:N,"ch1"))
alpha <- 0
beta <- 0
ksi <- c(1,1,1)
zeta <- 0.8
waves <- 4

simrec(nseeds, N, char, alpha, beta, ksi, zeta, waves)
}
