% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multr_unr.R
\name{urInfR}
\alias{urInfR}
\title{Inference for preference coefficients given observed recruitment chain, covariates, and prior values for preference coefficients.}
\usage{
urInfR(
  rdf,
  covnames,
  ma0,
  sa0,
  mb0,
  sb0,
  Kcalc,
  tallowed,
  ncoup = 3,
  restrict.local = FALSE,
  popadjmat = NULL,
  updlast = TRUE,
  usenetsize = FALSE
)
}
\arguments{
\item{rdf}{An \code{rds.data.frame} containing observed recruitment information.  Must also have a column for covariate(s).}

\item{covnames}{Vector of names of covariates in \code{rdf} to be used in the model.}

\item{ma0}{Vector of prior means for \eqn{alpha} and \eqn{ksi}.  Must be equal to \code{covnames+3} in length. The last threes values are for self-match.}

\item{sa0}{Matrix of prior variances for \eqn{alpha} and \eqn{ksi}.}

\item{mb0}{Vector of prior means for \eqn{beta} and \eqn{zeta}. Must be equal to \code{covnames+1} in length. The last value is for self-match.}

\item{sb0}{Matrix of prior variances for \eqn{beta} and \eqn{ksi}.}

\item{Kcalc}{Desired number of iterations (MCMC parameter). This is total, so Kcalc = burn + interval * noutiter.}

\item{tallowed}{Number of hours allowed on Hoffman.}

\item{ncoup}{Number of coupons recruiters had to distribute.}

\item{restrict.local}{Logical: use additional information about the underlying network? Makes computation faster. If TRUE, assumes that \code{netstage} takes three values: 1 indicates a recruitment tie (and thus a tie in the underlying network); 0 indicates no recruitment tie, but a tie does exist in the underlying network; NA indicates no tie in the underlying network, and thus no possibility of recruitment.}

\item{popadjmat}{Population adjacency matrix.  Must be supplied if \code{restrict.local=TRUE}.}

\item{updlast}{Update beta self values using information in last wave? Defaults to true.}

\item{usenetsize}{Should network size be used as a variable to approximate? If true, prior value must be reflected in ma0, sa0, mb0, and sb0.}
}
\value{
A list containing
 \item{alpha}{Value of \eqn{alpha} at each out-iteration.}
 \item{beta}{Value of \eqn{beta} at each out-iteration.}
 \item{rejcounter}{Number of rejections (due to swapping problem).}
 \item{earlystop}{Number of iterations after which Gibbs sampler was forced to abort for time constraints.}
}
\description{
Function that carries out inference step for preference coefficients for RDS.  Given an observed annotated adjacency matrix, covariates, prior values, and MCMC parameters, produces estimates for \eqn{alpha} and \eqn{beta}.
}
